model/requirements.txt
scikit-learn
pandas
mysql-connector-python